Rm register, improved login page with reqs
