+++ src/routes/(app)/+layout.svelte
@@ -10,7 +10,7 @@
           <div class="p-1 md:p-5 lg:p-10">
             <h4 class="h4">Contact information</h4>
             <br>
-            <span><i class="fa fa-map-pin"></i>&nbsp;&nbsp;<a href="https://maps.app.goo.gl/idEzGhBPDR5H2WSe8" class="hover:anchor">Ritarikatu 2 B, Helsinki</a></span><br>
+            <span><i class="fa fa-map-marker"></i>&nbsp;&nbsp;<a href="https://maps.app.goo.gl/idEzGhBPDR5H2WSe8" class="hover:anchor">Ritarikatu 2 B, Helsinki</a></span><br>
             <span><i class="fa fa-phone"></i>&nbsp;&nbsp;<a href="tel:+358 295 16001" class="hover:anchor">+358 295 16001</a></span><br>
             <span><i class="fa fa-envelope"></i>&nbsp;&nbsp;<a href="mailto:registry.vnk@gov.fi" class="hover:anchor">registry.vnk@gov.fi</a></span>
           </div>
